working capture of SQS context
